Automatically add the virsh IP to known hosts
During the generation of the bootstrap VM, ssh connection to the host
machine is opened. If the IP is not yet in the list of known hosts,
the installation blocks and waits for user input.
The fix queries the host and adds the ssh fingerprint(s) to the list of
known hosts, if it is not present already.

Change-Id: I3e4bbb80811ab880e7271970713047c5d4ccca0f
Signed-off-by: Martin Kulhavy <martin.kulhavy@nokia.com>