Improve and highlight logging to console (2)
xtrace produces a lot of output that can be hard to read. This commit
uses extra functions for formatted output (info and error),
so that they stand out in the log and make it easier for the user to
understand what is currently happening.

Building up on change #38909: https://gerrit.opnfv.org/gerrit/38909

Change-Id: I0ddf9e0d3a12a62ae45c56d33a8d48c335b618a0
Signed-off-by: Martin Kulhavy <martin.kulhavy@nokia.com>