Add a script to create proxies to dashboards
A script to create virtual hosts in Apache2 to proxy communication to
the web dashboards/consoles which might be on private networks. In case
of frequent access to these services, this approach is simpler than
using SSH tunneling each time.
Additionally, this script creates a customized homepage for the jumphost
with links to the dashboards and information about the credentials.

Change-Id: I2c9e17a95fa480095a4baba6bd990628bc73e053
Signed-off-by: Martin Kulhavy <martin.kulhavy@nokia.com>