Create new consolidated documentation for JOID
The current documentation contains many different versions, which
overlap and partly contradict each other. This makes it very hard for
people with no experience with JOID to deploy it for the first time.

This commit tries to create a new docs structure, merging information
currently spread across different files, while at the same time updating
the information and adding new hints based on the experience deploying
JOID in our lab.

The new docs structure is inspired by and tries to follow the docs
structure used for the APEX installer.

As a next step after finalizing this transformation, the other
overlapping docs directories shall be removed

Change-Id: I7075ea48565a7bd509321a532f1173cffd24cdbc
Signed-off-by: Martin Kulhavy <martin.kulhavy@nokia.com>